requirements-conda.txt
blast=2.6.0
bmtagger=3.101
bwa=0.7.17
cd-hit=4.6.8
cd-hit-auxtools=4.6.8
diamond=0.9.10
kmc=3.1.1rc1
gatk=3.8
kraken=1.0.0_fork3
krona=2.7
last=876
lbzip2=2.5
lz4-bin=131
fastqc=0.11.7
gap2seq=2.1
mafft=7.402
mummer4=4.0.0beta2
muscle=3.8.1551 
mvicuna=1.0
novoalign=3.07.00
parallel=20160622
picard=2.18.11
pigz=2.4
prinseq=0.20.4
samtools=1.9
snpeff=4.3.1t
spades=3.12.0
tbl2asn=25.6
trimmomatic=0.38
trinity=date.2011_11_26
unzip=6.0
vphaser2=2.0
# Python packages below
arrow=0.12.1
bedtools=2.27.1
biopython=1.72
future==0.16.0
matplotlib=1.5.3
pysam=0.15.0
pybedtools=0.7.10
PyYAML=5.1
